Cleaned up the DEinterface such that we’re using run_fvm_sim (instead of a huge interface file)
Other changes:
- Changed imports of Diffusion_Module to acousticDE
- headless_backend/HelperFunctions.py:
    - Improved plotting functionality
- Added 2x2x2 geo example.
