+++ simulation-backend/simulation_backend/headless_backend/HelperFunctions.py
@@ -49,9 +49,11 @@ def create_tmp_from_input(json_input_file):
     with open(json_input_file, "r") as json_input:
         data = json.load(json_input)
 
-    # Make the relative geometry file paths absolute
-    data["msh_path"] = find_input_file_in_subfolders(dirname, data["msh_path"])
+    # Make the relative geometry file path absolute
     data["geo_path"] = find_input_file_in_subfolders(dirname, data["geo_path"])
+    
+    # Generate .msh path from the .geo path
+    data["msh_path"] = os.path.splitext(data["geo_path"])[0] + ".msh"
 
     # Prepare the output file by copying the input into it. We need a separate output file to not overwrite the input file
     json_tmp_file = os.path.join(dirname, "MeasurementRoomTmp.json")
@@ -118,8 +120,13 @@ def plot_results(json_output):
     if json_output is not None:
         with open(json_output, "r") as json_file:
             result_container = json.load(json_file)
-
     plt.figure()
-    for i in range(4):
-        plt.plot(result_container["results"][0]["responses"][i]["receiverResultsUncorrected"])
+
+    if result_container["results"][0]["resultType"] == "DG":
+        for i in range(len(result_container["results"][0]["responses"])):
+            plt.plot(result_container["results"][0]["responses"][i]["receiverResultsUncorrected"])
+    elif result_container["results"][0]["resultType"] == "DE":
+        for i in range(len(result_container["results"][0]["responses"][0]["receiverResults"])):
+            plt.plot(result_container["results"][0]["responses"][0]["receiverResults"][i]["data"])
+        
     plt.show()
